+++ changelog.md
@@ -692,5 +692,6 @@ No changes. Erroneous version bump.
 ## Next
 
 * Fix: `controllers/_default/resource/new` do not load menu options for form fields relating to ancestors which are removed
+* Fix: `resourceJoin` default `edit` action do not load menu options for form field for both sides of relationship
 * `fieldsExclude` + `fieldsOnly` options on resource and resource join actions
 * Export `routes` object as `overlook.routes`