+++ changelog.md
@@ -695,4 +695,5 @@ No changes. Erroneous version bump.
 * Fix: `resourceJoin` default `edit` action do not load menu options for form field for both sides of relationship
 * `fieldsExclude` + `fieldsOnly` options on resource and resource join actions
 * Refactor default `root` controllers to use `fieldsExclude` option
+* View compilation errors include `routePath`
 * Export `routes` object as `overlook.routes`